Revert "Fix cinematic camera crash caused by cinematic distance accessor type"

+++ src/main/java/best/spaghetcodes/kira/mixins/EntityRendererAccessor.java
@@ -7,14 +7,14 @@ import org.spongepowered.asm.mixin.gen.Accessor;
 @Mixin(EntityRenderer.class)
 public interface EntityRendererAccessor {
     @Accessor("thirdPersonDistance")
-    double getThirdPersonDistance();
+    float getThirdPersonDistance();
 
     @Accessor("thirdPersonDistance")
-    void setThirdPersonDistance(double distance);
+    void setThirdPersonDistance(float distance);
 
     @Accessor("thirdPersonDistanceTemp")
-    double getThirdPersonDistanceTemp();
+    float getThirdPersonDistanceTemp();
 
     @Accessor("thirdPersonDistanceTemp")
-    void setThirdPersonDistanceTemp(double distance);
+    void setThirdPersonDistanceTemp(float distance);
 }
